refactor add button as component
